Split amp-timeago component files
